refactor: split db-store sync relations, tags, and write phases into modules
